Slightly broken
Restructured solvers + other code to extract A_matrices separately, and explicitly optimise for gamma if trapezoid, and x if rbf.
Changed loss function so gamma now updates.
Fit definitely worse for real data now. I am not sure why. Trapezoid runs for larger datasets but fit is bad if real data. rbf seems to run for smaller datasets but fit terrible for all. Noticed there is a problem in A matrices calculation, I think should be Toeplitz matrix but seems to not be.
Suspected multiple errors due to my ignorance of how to use jax and equinox packages + type declaration.